+++ CHANGELOG.md
@@ -7,6 +7,18 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased] - yyyy-mm-dd
 
+## [2.1.0] - 2022-09-03
+
+>**Important**:<br>In case you have not updated to Member Audit 2.x yet, please follow the special update instructions for updating to 2.0.0 first, before installing this update!
+
+### Added
+
+- Ability to create skill sets from imported skill plans
+
+### Changed
+
+- Error handling more user friendly when importing EFT fitting
+
 ## [2.0.1] - 2022-09-02
 
 >**Important**:<br>In case you have not updated to Member Audit 2.x yet, please follow the special update instructions for updating to 2.0.0 first, before installing this patch!